move and rename word list service under pipeline phases
