+++ src/core/server/stacks/editComment.ts
@@ -42,7 +42,7 @@ import {
   CreateCommentMediaInput,
 } from "coral-server/services/comments/media";
 import { processForModeration } from "coral-server/services/comments/pipeline";
-import { WordListService } from "coral-server/services/comments/pipeline/wordlistWorker/service";
+import { WordListService } from "coral-server/services/comments/pipeline/phases/wordList/service";
 import { AugmentedRedis } from "coral-server/services/redis";
 import { Request } from "coral-server/types/express";
